code review - rust syntax

+++ core/src/pow/types.rs
@@ -293,7 +293,7 @@ impl ProofOfWork {
 	/// Maximum unscaled difficulty this proof of work can achieve
 	pub fn to_unscaled_difficulty(&self) -> Difficulty {
 		// using scale = 1 gives "unscaled" value
-		Difficulty::from_num(self.proof.scaled_difficulty(1 as u64))
+		Difficulty::from_num(self.proof.scaled_difficulty(1u64))
 	}
 
 	/// The edge_bits used for the cuckoo cycle size on this proof
